created upload image - and favorite books

+++ routes/book.routes.js
@@ -4,6 +4,7 @@ const User = require('../models/User.model')
 const Book = require('../models/Book.model')
 const getBookReq = require('./book_functions/getBookReq')
 const verifyUserId = require('./helper_functions/verifyUserId.js')
+const uploadCloud = require('../config/cloudinary.config')
 
 const router = Router()
 
@@ -14,7 +15,7 @@ router.post('/', async (req, res) => {
 
   try {
   
-    const newBook = await Book.create({ name, author, userId, favoritedByUsers: userId })
+    const newBook = await Book.create({ name, author, createdByUser: userId, favoritedByUsers: userId })
 
     await User.findByIdAndUpdate(userId, { $push: { books: newBook._id }})
 
@@ -23,6 +24,26 @@ router.post('/', async (req, res) => {
   } catch (error) {
     
     res.status(500).json({ place: 'Error trying to create a new book', error: error.message })
+
+  }
+
+})
+
+// Upload a Book Image
+router.put('/image/:bookId', uploadCloud.single('image') ,async (req, res) => {
+
+  const { bookId } = getBookReq(req)
+  const { path } = req.file
+
+  try {
+    
+    const updatedBook = await Book.findByIdAndUpdate(bookId, { imageUrl: path }, { new: true })
+    
+    res.status(200).json(updatedBook)
+
+  } catch (error) {
+
+    res.status(error.status || 500).json({ place: "Error trying upload book image", error: error.message })
 
   }
 
@@ -45,6 +66,25 @@ router.get('/', async (req, res) => {
 
 })
 
+// Get favorite books by the user
+router.get('/favorites', async (req, res) => {
+
+  const { userId } = getBookReq(req)
+
+  try {
+    
+    const favoritedBooks = await Book.find({ favoritedByUsers: userId })
+
+    res.status(200).json(favoritedBooks)
+
+  } catch (error) {
+    
+    res.status(error.status || 500).json({ place: 'Error trying to get favorite books', error: error.message })
+
+  }
+
+})
+
 // Get one BookById
 router.get('/:bookId', async (req, res) => {
 
@@ -57,6 +97,7 @@ router.get('/:bookId', async (req, res) => {
     res.status(200).json(findBook)
 
   } catch (error) {
+
     res.status(500).json({ place: 'Error trying to get one book', error: error.message })
 
   }
@@ -70,7 +111,7 @@ router.delete('/delete/:bookId', async (req, res) => {
 
   try {
 
-    const deleted = await Book.findOneAndDelete({ _id: bookId, userId })
+    const deleted = await Book.findOneAndDelete({ _id: bookId, createdByUser: userId })
 
     verifyUserId(deleted, "You can't delete a book created by another user")
 
@@ -85,5 +126,6 @@ router.delete('/delete/:bookId', async (req, res) => {
   }
 
 })
+
 
 module.exports = router
